♻️ [refactor]: #452 Phase 3b: lexer_is_eof.rs を分離
is_eof の 2 テストを `src/lexer/tests/lexer_is_eof.rs` に分離。
tests.rs から Phase 3 ヘッダコメントを削除（Phase 3a + 3b で全移動完了）。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_is_eof.rs: 18 行 (< 800)

Refs #452
